fix:     1、日志分页查询bug     2、删除AdminUserService.findById方法

+++ src/main/java/com/am/server/common/base/pojo/vo/PageVO.java
@@ -49,8 +49,8 @@ public class PageVO<T> implements Serializable {
      * @return int
      */
     @ApiModelProperty(value = "当前页", example = "200")
-    public int getTotalPage() {
-        return (total % pageSize == 0 ? (total / pageSize) : (total / pageSize + 1));
+    public Integer getTotalPage() {
+        return Math.toIntExact((total % pageSize == 0 ? (total / pageSize) : (total / pageSize + 1)));
     }
 
     /**
